Finished commenting all the code I wrote.

+++ lib/Organization/add_urgentcase_screen.dart
@@ -8,7 +8,7 @@ import 'package:flutter/services.dart';
 import 'package:modal_progress_hud_nsn/modal_progress_hud_nsn.dart';
 import 'package:get/get.dart';
 
-
+// Add urgent case form
 class AddUrgentCaseForm extends StatefulWidget {
   static const id = 'urgentcase_form_screen';
   AddUrgentCaseForm({Key? key}) : super(key: key);
@@ -36,7 +36,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
   var category = [];
   int urgentCaseTimeLimit=0;
 
-
+// Get category
   _getCampaign() async {
     var ret = await firestore
         .collection('CharityCategories')
@@ -47,7 +47,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
 
     setState(() {});
   }
-
+// Get admin time limit from firebase
   _getTimeLimit() async {
     var ret = await firestore.collection('AdminRestrictions').where('id',isEqualTo: 'CharityDurationLimits').get();
 
@@ -67,7 +67,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
     loggedInUser = _auth.currentUser;
   }
 
-
+//Add urgent case to firebase
   Future<void> addUrgentCase(String category, String description, double goalAmount,
       String title, String endDateController) async {
     try {
@@ -93,7 +93,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
     }
   }
 
-
+// UI for for urgent case
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -119,6 +119,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                         const SizedBox(
                           height: 15.0,
                         ),
+                        // Add urgent case from UI
                          Center(
                           child: Text(
                             'add_urgent_case'.tr,
@@ -136,6 +137,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                             style: TextStyle(color: Colors.red),
                           ),
                         ),
+                        // Error message please enter title
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: TextFormField(
@@ -151,6 +153,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                 return null;
                               }
                             },
+                            //// UI for title enter field
                             textAlign: TextAlign.center,
                             decoration: InputDecoration(
                                 label: Center(
@@ -176,6 +179,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                 )),
                           ),
                         ),
+                        // Error message please enter description
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: TextFormField(
@@ -190,6 +194,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                 return null;
                               }
                             },
+                            //description field ui
                             textAlign: TextAlign.center,
                             decoration: InputDecoration(
                                 label: Center(
@@ -215,6 +220,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                 )),
                           ),
                         ),
+                        //Please enter goal amount error message or enter valid goal amount
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: TextFormField(
@@ -230,6 +236,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                 return null;
                               }
                             },
+                            //Goal amount UI
                             textAlign: TextAlign.center,
                             decoration: InputDecoration(
                                 label: Center(
@@ -255,6 +262,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                 )),
                           ),
                         ),
+                        //Please valid enter end date error
                         Padding(
                             padding: const EdgeInsets.all(8.0),
                             child: TextFormField(
@@ -304,7 +312,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                 endDateController.text =
                                     date.toString().substring(0, 10);
                               },)),
-
+                        // Category field UI
                         Padding(
                             padding: const EdgeInsets.all(8.0),
                             child: DropdownButtonFormField <String>(
@@ -347,6 +355,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                   ? 'please_fill_in_the_category'.tr : null,
                             )
                         ),
+                        //Urgent case note for approval
                         Padding(
                           padding: const EdgeInsets.all(8.0),
                           child: Container(
@@ -368,6 +377,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                             ),
                           ),
                         ),
+                        //Submit button and navigate to the dashboard
                         Padding(
                           padding: const EdgeInsets.symmetric(
                               vertical: 20.0, horizontal: 5.0),
@@ -389,7 +399,7 @@ class _AddUrgentCaseFormState extends State<AddUrgentCaseForm> {
                                   });
                                   addUrgentCase(categoryController.text,
                                       descriptionController.text,
-                                      double.parse(goalAmountController.text),
+                                      double.parse(goalAmountController.text).toDouble(),
                                       titleController.text,
                                       endDateController.text);
                                   Navigator.of(context).popUntil(ModalRoute.withName(OrganizationDashboard.id));
